Handler events in modal windows + add styles to likes and comments
