+++ src/controller/mainController.js
@@ -1,12 +1,12 @@
 import signInCtrl from './signInController.js';
 import signUpCtrl from './signUpController.js';
-import homeCtrl from './homeController.js';
+import accountCtrl from './accountController.js';
 // import postCtrl from './postController.js';
 
 const controllers = {
   signInCtrl,
   signUpCtrl,
-  homeCtrl,
+  accountCtrl,
   // postCtrl,
 };
