+++ src/controller/postController.js
@@ -1,9 +1,28 @@
 /* eslint-disable no-console */
 import {
-  createPostBD, deletePostBD, getPostBD, updatePostBD, getAllPostsBD,
+  createPostBD,
 } from '../model/post.model.js';
 import { uploadImage } from '../model/storage-post.js';
-import { post } from '../view/post.js';
 
-
+export const createPost = (user, text, images, statePrivacity) => {
+  const postObj = {
+    textContent: text,
+    imageContent: '',
+    likes: [],
+    comments: [],
+    privacity: statePrivacity,
+    date: firebase.firestore.FieldValue.serverTimestamp(),
+    nameUser: user.displayName,
+    idUser: user.uid,
+    photoUser: user.photoURL,
+  };
+  if (images[0]) {
+    uploadImage(images[0])
+      .then((url) => {
+        postObj.imageContent = url;
+        createPostBD(postObj);
+      });
+  } else {
+    createPostBD(postObj);
+  }
 };
