+++ src/index.html
@@ -23,8 +23,7 @@
 
     </section>
 
-    <div class="bg-modal">      
-      
+    <div class="bg-modal">            
     </div>
   </main>
