+++ src/styles/css/style.css
@@ -33,6 +33,10 @@
   right: 0;
 }
 
+.liked {
+  color: #f16464 !important;
+}
+
 .modal .modal-footer button, .my-self .info-user button {
   font-weight: bold;
   margin: 0.2em;
@@ -75,6 +79,10 @@ section#container {
   display: flex;
   width: 100%;
   height: 100%;
+  overflow-y: auto;
+}
+section#container::-webkit-scrollbar {
+  display: none;
 }
 
 .bg-modal {
@@ -131,6 +139,10 @@ input, button, textarea {
   right: 0;
 }
 
+.liked {
+  color: #f16464 !important;
+}
+
 .modal .modal-footer button, .my-self .info-user button {
   font-weight: bold;
   margin: 0.2em;
@@ -162,7 +174,8 @@ input, button, textarea {
   display: flex;
   justify-content: center;
   width: 100%;
-  height: 100%;
+  height: max-content;
+  min-height: 100%;
 }
 
 .screen-background .form-container button, .screen-background .form-container form .form-control input {
@@ -198,6 +211,10 @@ input, button, textarea {
 .right {
   position: relative;
   right: 0;
+}
+
+.liked {
+  color: #f16464 !important;
 }
 
 .modal .modal-footer button, .my-self .info-user button {
@@ -356,6 +373,10 @@ header nav ul li {
   right: 0;
 }
 
+.liked {
+  color: #f16464 !important;
+}
+
 .modal .modal-footer button, .my-self .info-user button {
   font-weight: bold;
   margin: 0.2em;
@@ -388,6 +409,10 @@ header nav ul li {
   margin-right: 0.5em;
   border-radius: 50px;
   display: block;
+}
+
+.enabled {
+  background-color: #0CA385 !important;
 }
 
 .posts-container .post .post-header div.name-date-post span {
@@ -511,6 +536,7 @@ header nav ul li {
   background-color: #f1f5f6;
 }
 .posts-container .post .post-comments {
+  transition: all 0.9s;
   display: none;
 }
 .posts-container .post .post-comments .create-comment-container {
@@ -545,10 +571,7 @@ section.content {
   background: #e6f0ed;
   padding: 60px 0.5em;
   width: 98%;
-  overflow-y: auto;
-}
-section.content::-webkit-scrollbar {
-  display: none;
+  height: max-content;
 }
 section.content section {
   margin: 0.5em 0;
@@ -634,12 +657,13 @@ section.content .create-post-container .btn-container .opts:hover {
 section.content .create-post-container .btn-container .post-btn {
   position: relative;
   right: 0;
+  border: none;
+  font-weight: bold;
   background: rgba(12, 163, 133, 0.514);
   color: white;
 }
-section.content .create-post-container .btn-container .post-btn:hover {
+section.content .create-post-container .btn-container .post-btn:enabled {
   cursor: pointer;
-  background-color: #0CA385;
 }
 
 .div-img {
@@ -650,10 +674,10 @@ section.content .create-post-container .btn-container .post-btn:hover {
   margin-right: 0.3em;
 }
 .div-img:hover {
-  background-color: rgba(27, 26, 26, 0.349);
+  background-color: rgba(27, 26, 26, 0.137);
 }
 .div-img:hover i {
-  color: #ecf3f1;
+  color: #707472;
   font-weight: bolder;
 }
 .div-img i {
@@ -807,6 +831,10 @@ section.content .create-post-container .btn-container .post-btn:hover {
 .right {
   position: relative;
   right: 0;
+}
+
+.liked {
+  color: #f16464 !important;
 }
 
 .modal .modal-footer button, .my-self .info-user button {
@@ -1094,6 +1122,10 @@ section.content .create-post-container .btn-container .post-btn:hover {
   right: 0;
 }
 
+.liked {
+  color: #f16464 !important;
+}
+
 .modal .modal-footer button, .my-self .info-user button {
   font-weight: bold;
   margin: 0.2em;
@@ -1125,14 +1157,17 @@ section.content .create-post-container .btn-container .post-btn:hover {
   border-radius: 5px;
   background-color: white;
   border: 1px solid rgba(153, 153, 153, 0.3);
-  width: 90%;
+  width: 80%;
+  display: flex;
+  flex-direction: column;
+  min-width: 250px;
 }
 .modal .modal-header, .modal .modal-footer {
   padding: 0.5em;
   background: #F2F2F2;
+  width: auto;
   font-weight: bold;
   color: #080827;
-  width: 100%;
 }
 .modal .modal-body {
   padding: 1em;
@@ -1143,7 +1178,7 @@ section.content .create-post-container .btn-container .post-btn:hover {
   justify-content: flex-end;
 }
 .modal-delete {
-  max-width: 400px;
+  max-width: 500px;
 }
 
 .modal-edit {
