+++ src/styles/sass/layouts/_modal.scss
@@ -7,13 +7,16 @@
   border-radius: 5px;
   background-color: white;
   border: 1px solid $color-border;    
-  width: 90%;
+  width: 80%;
+  display: flex;
+  flex-direction: column;
+  min-width: 250px;
   .modal-header,.modal-footer{
     padding: .5em;
     background: $post-color;
+    width: auto;
     font-weight: bold;
     color: rgb(8, 8, 39);
-    width: 100%;
   }
   .modal-body{
     padding: 1em;
@@ -28,7 +31,7 @@
   }
 }
 .modal-delete{
-  max-width: 400px;
+  max-width: 500px;
 }
 .modal-edit{
   max-width: 600px;
@@ -39,9 +42,7 @@
   .modal-header{
     display: flex;
     justify-content: space-between;
-    i{
-      @extend %grow;
-    }
+    i{ @extend %grow; }
   }
   .modal-body{
     display: flex;      
@@ -49,6 +50,7 @@
       @include square-size(2em);
     }
     div[contenteditable]{
+      // autofocus:"autofocus";
       max-height: 200px;
       text-align: left;
       overflow: hidden;
@@ -66,4 +68,3 @@
     position: absolute;
   }
 }
-
