+++ src/styles/sass/pages/_home.scss
@@ -5,7 +5,8 @@
   display: flex;
   justify-content: center;
   width: 100%;
-  height: 100%;
+  height: max-content;
+  min-height: 100%;
 }
