+++ src/styles/sass/style.scss
@@ -15,6 +15,11 @@ section#container{
   display: flex;    
   width: 100%;
   height: 100%;
+  overflow-y: auto;  
+  &::-webkit-scrollbar { 
+    display: none;
+  }
+ 
 }
 .bg-modal{
   position: absolute;
@@ -25,7 +30,6 @@ section#container{
   width: 100%;
   height: 100%;
   background-color: rgba(0, 0, 0, 0.7);
-  
 }
