+++ src/view/index.js
@@ -1,6 +1,6 @@
 import signInView from './signIn.js';
 import signUpView from './signUp.js';
-import homeView from './home.js';
+import accountView from './account.js';
 // import postView from './post.js';
 import sendEmailView from './sendEmail.js';
 
@@ -8,7 +8,7 @@ const views = {
   signUpView,
   sendEmailView,
   signInView,
-  homeView,
+  accountView,
   // postView,
 };
